Refactor sightings data model adding sighting's jellyfishes size and abundance as separate models.
Before this commit, jellyfishes where created as variables (presence, size and abundance), and related to sightings among with an certain value.

Now Jellyfish is a model by itself, as the possible values of size and abundance. A sighting (Sight model) is related to several jellyfishes, among with a FK to their size and abundance.